modules/tool-recipes/pyproject.toml
[project]
name = "amplifier-module-tool-recipes"
version = "0.1.0"
description = "Tool module for executing multi-step AI agent recipes"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Microsoft", email = "amplifier@microsoft.com" }
]

dependencies = [
    "pyyaml>=6.0",
]

[project.entry-points."amplifier.modules"]
tool-recipes = "amplifier_module_tool_recipes:mount"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "amplifier-core",
]

[tool.uv.sources]
amplifier-core = { path = "../../../amplifier-core", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["amplifier_module_tool_recipes"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
